requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.4.1
dill==0.3.8
distro==1.9.0
fastapi==0.115.12
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
googletrans==4.0.0rc1
gpt4all==2.8.2
greenlet==3.2.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.29.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.3
openai==1.66.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.2
xxhash==3.5.0
yarl==1.18.3